Proposal for performance metrics
In addition to the number of recomputations for a selector it would be nice to know the total computation time for a given selector. This way a very expensive selector that gets called infrequently can still be identified.

I'm not very familiar with javascript so i'm not sure how accurate this code change is.